cnn-pytorch/requirements.txt
Flask
torch
torchvision
Pillow

